more explicit cleaning

+++ clean.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 # restore to pre-setup state
 
-rm -rf "$HOME/.bashrc" "$HOME/.zshrc" "$HOME/.profile"
+read -r -p 'This removes a lot of files. Are you sure? ' yn
+if [[ "$yn" != 'y' ]]; then
+    exit 0
+fi
+
+rm -rf "$HOME/.bashrc" "$HOME/.zshrc" "$HOME/.profile" "$HOME/.condarc"
 rm -rf "$HOME/.editorconfig" "$HOME/.tmux.conf"
 
 read -r -p 'Delete scripts? This includes extra scripts that may not be saved ' yn
